requirements.txt
requests==2.32.5
beautifulsoup4==4.14.2
openai==0.28.1
python-dotenv==1.0.1
